refactor: AutoBoardCore FreeBoardCore
