+++ frontend/src/app/pages/DashBoardPage/pages/Board/index.tsx
@@ -33,8 +33,8 @@ import { FullScreenPanel } from '../../components/FullScreenPanel';
 import TitleHeader from '../../components/TitleHeader';
 import { editDashBoardInfoActions } from '../BoardEditor/slice';
 import { clearEditBoardState } from '../BoardEditor/slice/actions/actions';
-import AutoBoardCore from './AutoDashboard/AutoBoardCore';
-import FreeBoardCore from './FreeDashboard/FreeBoardCore';
+import { AutoBoardCore } from './AutoDashboard/AutoBoardCore';
+import { FreeBoardCore } from './FreeDashboard/FreeBoardCore';
 import { boardActions } from './slice';
 import { makeSelectBoardConfigById } from './slice/selector';
 import { fetchBoardDetail } from './slice/thunk';
